python-encryption/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
exceptiongroup==1.3.0
fastapi==0.121.0
h11==0.16.0
idna==3.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
